requirements.txt
spotipy
yt-dlp
python-dotenv
mutagen
